# Reduce register usage of fused adam(w)