--- a/aten/src/ATen/native/cuda/fused_adam_amsgrad_impl.cu
+++ b/aten/src/ATen/native/cuda/fused_adam_amsgrad_impl.cu
@@ -45,7 +45,7 @@ void _fused_adam_amsgrad_cuda_impl_(
         multi_tensor_apply_for_fused_optimizer<5>(
             tensor_lists,
             state_steps,
-            FusedAdamMathFunctor<scalar_t, 5>(),
+            FusedAdamMathFunctor<scalar_t, 5, ADAM_MODE::ORIGINAL>(),
             lr_ptr, // unused
             lr,
             beta1,
@@ -55,8 +55,7 @@ void _fused_adam_amsgrad_cuda_impl_(
             maximize,
             /* amsgrad */ true,
             grad_scale_ptr,
-            found_inf_ptr,
-            ADAM_MODE::ORIGINAL);
+            found_inf_ptr);
       });
 }
 
@@ -98,7 +97,7 @@ void _fused_adam_amsgrad_cuda_impl_(
         multi_tensor_apply_for_fused_optimizer<5>(
             tensor_lists,
             state_steps,
-            FusedAdamMathFunctor<scalar_t, 5>(),
+            FusedAdamMathFunctor<scalar_t, 5, ADAM_MODE::ORIGINAL>(),
             lr_ptr,
             1.0, // unused
             beta1,
@@ -108,8 +107,7 @@ void _fused_adam_amsgrad_cuda_impl_(
             maximize,
             /* amsgrad */ true,
             grad_scale_ptr,
-            found_inf_ptr,
-            ADAM_MODE::ORIGINAL);
+            found_inf_ptr);
       });
 }