Part of #117872

| branch | cpu time avg (ms) | cuda time avg (ms) |
|--------|--------------|---------------|
| [main](eebe7e1d37f1baa995c694d540cc2fc98884fa18) | 13.430 | 144.117 |
| pr                                               | 13.371 | 49.655  | 

Used torch profiler to measure the avg perf or 20 iterations. 
Model is openlm-research/open_llama_7b_v2 (script is [here](https://gist.github.com/crcrpar/ca951d4e7f3e1c771d502135b798f0d1)).

---

PR
```
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------
                                                   Name    Self CPU %      Self CPU   CPU total %     CPU total  CPU time avg     Self CUDA   Self CUDA %    CUDA total  CUDA time avg       CPU Mem  Self CPU Mem      CUDA Mem  Self CUDA Mem    # of Calls
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------
                                          ProfilerStep*         0.00%       0.000us         0.00%       0.000us       0.000us        5.789s        46.42%        5.789s     289.456ms           0 b           0 b           0 b           0 b            20
                                          ProfilerStep*        36.02%        3.119s        67.19%        5.819s     290.958ms       0.000us         0.00%        2.586s     129.276ms      48.00 Kb      -1.47 Mb           0 b    -504.23 Gb            20
                                               aten::mm         2.57%     222.681ms         8.80%     762.415ms      56.475us        2.501s        20.05%        2.501s     185.255us           0 b           0 b     441.39 Gb     441.39 Gb         13500
       autograd::engine::evaluate_function: MmBackward0         0.10%       8.600ms         8.17%     707.935ms     157.319us       0.000us         0.00%        1.625s     361.098us           0 b           0 b     198.65 Gb    -135.03 Gb          4500
                                            MmBackward0         0.39%      33.896ms         7.99%     692.035ms     153.786us       0.000us         0.00%        1.601s     355.710us           0 b           0 b     330.84 Gb    -248.00 Mb          4500
                              Optimizer.step#AdamW.step         0.00%       0.000us         0.00%       0.000us       0.000us        1.007s         8.07%        1.007s      50.329ms           0 b           0 b           0 b           0 b            20
                                             AdamW.step         0.01%     837.000us         3.36%     290.610ms      14.530ms       0.000us         0.00%     993.235ms      49.662ms           0 b           0 b           0 b           0 b            20
                              Optimizer.step#AdamW.step         0.22%      18.825ms         3.35%     289.773ms      14.489ms       0.000us         0.00%     993.235ms      49.662ms           0 b           0 b           0 b           0 b            20
                                    aten::_fused_adamw_         0.12%      10.823ms         3.09%     267.428ms      13.371ms     993.095ms         7.96%     993.095ms      49.655ms           0 b           0 b           0 b           0 b            20
void at::native::(anonymous namespace)::multi_tensor...         0.00%       0.000us         0.00%       0.000us       0.000us     993.095ms         7.96%     993.095ms     154.207us           0 b           0 b           0 b           0 b          6440
                                           aten::matmul         0.19%      16.140ms         1.73%     149.869ms      33.304us       0.000us         0.00%     876.000ms     194.667us           0 b           0 b     107.46 Gb           0 b          4500
sm90_xmma_gemm_bf16bf16_bf16f32_f32_nt_n_tilesize128...         0.00%       0.000us         0.00%       0.000us       0.000us     835.374ms         6.70%     835.374ms     185.639us           0 b           0 b           0 b           0 b          4500
                                           aten::linear         0.27%      23.268ms         1.97%     170.227ms      37.828us       0.000us         0.00%     776.278ms     172.506us           0 b           0 b     107.46 Gb      12.17 Gb          4500
sm90_xmma_gemm_bf16bf16_bf16f32_f32_tn_n_tilesize128...         0.00%       0.000us         0.00%       0.000us       0.000us     707.074ms         5.67%     707.074ms     183.180us           0 b           0 b           0 b           0 b          3860
                                              aten::mul         1.31%     113.614ms         5.14%     445.405ms      22.125us     552.421ms         4.43%     552.780ms      27.459us     256.32 Kb     256.21 Kb     420.38 Gb     419.88 Gb         20131
sm90_xmma_gemm_bf16bf16_bf16f32_f32_nn_n_tilesize128...         0.00%       0.000us         0.00%       0.000us       0.000us     442.209ms         3.55%     442.209ms     138.190us           0 b           0 b           0 b           0 b          3200
      autograd::engine::evaluate_function: MulBackward0         0.25%      21.336ms         5.00%     432.976ms      74.651us       0.000us         0.00%     398.627ms      68.729us           0 b           0 b     -45.71 Gb    -252.76 Gb          5800
                                             aten::add_         0.37%      31.975ms         7.19%     622.433ms      53.658us     391.957ms         3.14%     391.957ms      33.789us           0 b           0 b      -4.35 Gb      -4.35 Gb         11600
sm90_xmma_gemm_bf16bf16_bf16f32_f32_nn_n_tilesize256...         0.00%       0.000us         0.00%       0.000us       0.000us     345.037ms         2.77%     345.037ms     265.413us           0 b           0 b           0 b           0 b          1300
                                            aten::copy_         0.41%      35.727ms        20.62%        1.786s     146.503us     342.386ms         2.75%     342.386ms      28.092us      48.00 Kb      48.00 Kb     -56.00 Mb     -56.00 Mb         12188
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------
Self CPU time total: 8.661s
Self CUDA time total: 12.472s
```

main
```
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------
                                                   Name    Self CPU %      Self CPU   CPU total %     CPU total  CPU time avg     Self CUDA   Self CUDA %    CUDA total  CUDA time avg       CPU Mem  Self CPU Mem      CUDA Mem  Self CUDA Mem    # of Calls
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------
                                          ProfilerStep*         0.00%       0.000us         0.00%       0.000us       0.000us        7.671s        42.31%        7.671s     383.529ms           0 b           0 b           0 b           0 b            20
                                          ProfilerStep*        28.85%        3.050s        72.83%        7.700s     385.009ms       0.000us         0.00%        4.474s     223.678ms      48.00 Kb      -1.48 Mb           0 b    -504.45 Gb            20
                              Optimizer.step#AdamW.step         0.00%       0.000us         0.00%       0.000us       0.000us        2.896s        15.97%        2.896s     144.787ms           0 b           0 b           0 b           0 b            20
                                             AdamW.step         0.01%     819.000us         2.75%     291.024ms      14.551ms       0.000us         0.00%        2.882s     144.125ms           0 b           0 b           0 b           0 b            20
                              Optimizer.step#AdamW.step         0.17%      18.291ms         2.74%     290.205ms      14.510ms       0.000us         0.00%        2.882s     144.125ms           0 b           0 b           0 b           0 b            20
                                    aten::_fused_adamw_         0.10%      10.893ms         2.54%     268.602ms      13.430ms        2.882s        15.90%        2.882s     144.117ms           0 b           0 b           0 b           0 b            20
void at::native::(anonymous namespace)::multi_tensor...         0.00%       0.000us         0.00%       0.000us       0.000us        2.882s        15.90%        2.882s     447.570us           0 b           0 b           0 b           0 b          6440
                                               aten::mm         2.05%     217.136ms         7.21%     762.211ms      56.460us        2.499s        13.78%        2.499s     185.075us           0 b           0 b     441.37 Gb     441.37 Gb         13500
       autograd::engine::evaluate_function: MmBackward0         0.07%       7.179ms         6.77%     715.673ms     159.038us       0.000us         0.00%        1.624s     360.812us           0 b           0 b     198.65 Gb    -134.64 Gb          4500
                                            MmBackward0         0.32%      34.257ms         6.62%     700.088ms     155.575us       0.000us         0.00%        1.600s     355.460us           0 b           0 b     330.59 Gb    -628.00 Mb          4500
                                           aten::matmul         0.15%      15.892ms         1.32%     139.597ms      31.022us       0.000us         0.00%     874.861ms     194.414us           0 b           0 b     107.46 Gb           0 b          4500
sm90_xmma_gemm_bf16bf16_bf16f32_f32_nt_n_tilesize128...         0.00%       0.000us         0.00%       0.000us       0.000us     834.631ms         4.60%     834.631ms     185.474us           0 b           0 b           0 b           0 b          4500
                                           aten::linear         0.21%      22.460ms         1.51%     159.620ms      35.471us       0.000us         0.00%     774.772ms     172.172us           0 b           0 b     107.46 Gb      11.88 Gb          4500
sm90_xmma_gemm_bf16bf16_bf16f32_f32_tn_n_tilesize128...         0.00%       0.000us         0.00%       0.000us       0.000us     705.996ms         3.89%     705.996ms     182.901us           0 b           0 b           0 b           0 b          3860
                                              aten::mul         1.06%     112.529ms         4.28%     452.473ms      22.488us     552.242ms         3.05%     552.266ms      27.447us     255.90 Kb     255.88 Kb     413.93 Gb     413.90 Gb         20121
sm90_xmma_gemm_bf16bf16_bf16f32_f32_nn_n_tilesize128...         0.00%       0.000us         0.00%       0.000us       0.000us     441.514ms         2.44%     441.514ms     137.973us           0 b           0 b           0 b           0 b          3200
      autograd::engine::evaluate_function: MulBackward0         0.19%      20.517ms         4.18%     442.189ms      76.239us       0.000us         0.00%     398.552ms      68.716us           0 b           0 b     -45.57 Gb    -251.17 Gb          5800
                                             aten::add_         0.30%      31.703ms         6.01%     635.030ms      54.744us     391.897ms         2.16%     391.897ms      33.784us           0 b           0 b      -5.71 Gb      -5.71 Gb         11600
sm90_xmma_gemm_bf16bf16_bf16f32_f32_nn_n_tilesize256...         0.00%       0.000us         0.00%       0.000us       0.000us     344.972ms         1.90%     344.972ms     265.363us           0 b           0 b           0 b           0 b          1300
                                            aten::copy_         0.33%      34.415ms        34.75%        3.674s     301.437us     342.661ms         1.89%     342.661ms      28.115us      80.00 Kb      80.00 Kb    -240.00 Mb    -240.00 Mb         12188
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------
Self CPU time total: 10.574s
Self CUDA time total: 18.129s
```