@janeyx99 2024-01-30
I'm also curious what numbers you get for APEX fused adam in comparison, if you have those!