@crcrpar 2024-01-31
> I'm also curious what numbers you get for APEX fused adam in comparison, if you have those!

apex fused adam(w) seems to consume more memory partly because their moving averages of grads and squared grads are always in fp32, leading to OOM in the script