@janeyx99 2024-02-05
I see a general lift across the eager benchmarks:
![image](https://github.com/pytorch/pytorch/assets/31798555/afb5f4d7-2cf4-4718-8fe0-1214139665f3)